Refactor Spotify UI optimistic state and volume sync
- Extracted volume state reducer logic and optimistic syncing into `useSpotifyVolume` to DRY up `SpotifyControls` and `SpotifyDisplay`
- Removed `useVolumePreference` and `clampVolume` from hooks, inlining essential logic to `AudioContext` and centralizing `clampVolume` in `audioManager`
- Replaced deep cloning with shallow copy in `SpotifyPlayerManager.executeOptimistic` to prevent performance drops during rapid volume slider movement
- Consolidated `VolumeSlider` styling to scale linearly based on a single `SCALE_FACTOR`
- Simplified UI device selection boilerplate and fixed `useEffect` dependency reference bugs in `SpotifyControls`
- Addressed code review feedback and cleared trailing agent tool artifacts from repository
